requirements.txt
discord.py-self
pygame
python-dotenv
twilio
